still not working

+++ src/AS5247Spi/AS5247Spi.h
@@ -46,7 +46,7 @@ class AS5247Spi {
     public:
     
         AS5247Spi(uint8_t chipSelectPin);
-        void writeData(uint32_t command, uint32_t value);
+        void writeData(uint16_t command, uint16_t value);
         uint32_t readData(uint32_t command, uint32_t nopCommand);
         uint16_t readData16(uint16_t command, uint16_t nopCommand);
